requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
passlib
python-jose[cryptography]
python-multipart
pydantic-settings
PyJWT